+++ src/plugins/ipaddr/CMakeLists.txt
@@ -4,5 +4,6 @@ add_plugin (ipaddr
 	SOURCES
 		ipaddr.h
 		ipaddr.c
+		test_ipaddr.h
 	ADD_TEST
 	)
